+++ paylogic/lookups.py
@@ -47,11 +47,11 @@ class CaseAssignedView(Select2View):
         return (NO_ERR_RESP, False, possible_assignees)  # Any error response, Has more results, options list
 
 
-class CaseTagsView(Select2View):
+class TagsView(Select2View):
 
-    """Fetch case tags."""
+    """Fetch available tags."""
 
     def get_results(self, request, term, page, context):
-        case_id = self.kwargs['case_id']
+        case_id = self.kwargs.get('case_id')
         tags = views.get_fogbugz_tags(request, case_id)
         return (NO_ERR_RESP, False, tags)  # Any error response, Has more results, options list